feat : 장바구니에 들어가 있는 상품 조회

+++ src/main/java/com/example/lp/product/dto/request/ProductSkuRequest.java
@@ -3,6 +3,6 @@ package com.example.lp.product.dto.request;
 public record ProductSkuRequest(String skuCode,
                                 String color,
                                 String size,
-                                int stock,
-                                int price,
+                                Long stock,
+                                Long price,
                                 boolean inActive) {}
